removed face_recognition module, detecting face using opencv
