+++ terraform/segment_modifier_lambda/lambda/requirements.txt
@@ -1,24 +1,3 @@
-aiohttp==3.9.5
-aiosignal==1.3.1
-attrs==23.2.0
-boto3==1.34.95
-botocore==1.34.95
-certifi==2024.2.2
-charset-normalizer==3.3.2
-click==8.1.7
-dlib==19.24.4
-face-recognition==1.3.0
-face-recognition-models==0.3.0
-frozenlist==1.4.1
-idna==3.7
-jmespath==1.0.1
-multidict==6.0.5
-numpy==1.26.4
-opencv-python==4.9.0.80
-pillow==10.3.0
-python-dateutil==2.9.0.post0
 requests==2.31.0
-s3transfer==0.10.1
-six==1.16.0
-urllib3==2.2.1
-yarl==1.9.4
+aiohttp==3.9.5
+opencv-python==4.6.0.66
